Code/requirements.txt
appnope==0.1.0
autocorrect==0.3.0
bcrypt==3.1.6
bleach==2.1.3
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-heroku==0.3.1
entrypoints==0.2.3
gunicorn==19.9.0
html5lib==1.0b10
idna==2.8
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.8
oauthlib==3.0.1
olefile==0.44
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pipenv==2018.11.26
prompt-toolkit==1.0.15
psycopg2==2.7.7
ptyprocess==0.5.2
pycparser==2.19
Pygments==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
qtconsole==4.3.1
requests==2.21.0
requests-oauthlib==1.2.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.4.0
virtualenv-clone==0.5.1
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.0.8
